- vycisten interface EntityMgr - fix StringKeye, aby sel buildnout

+++ src/Utils/StringKey.h
@@ -57,7 +57,7 @@ namespace Utils
 		struct StringKeyData
 		{
 			StringKeyData(string refString): mRefCount(0), mRefString(refString) {}
-			uint    mRefCount;
+			uint8   mRefCount;
 			string  mRefString;
 		};
